Clean up config code
Move util functions out of the catch-all config file.
Split loadConfig out into its own file.
