+++ packages/apollo-language-server/src/config/index.ts
@@ -1,1 +1,3 @@
+export * from "./utils";
 export * from "./config";
+export * from "./loadConfig";
